adding friends search and bar
