+++ client/src/utils/API.js
@@ -23,10 +23,10 @@ export default {
 		return axios.get(`/api/trailers/${title}/${year}`)
 	},
 	addFriend: function(user, friend) {
-		console.log("------Add Friend------");
-		console.log(user);
-		console.log(friend);
-		console.log("------Add Friend------");
+		// console.log("------Add Friend------");
+		// console.log(user);
+		// console.log(friend);
+		// console.log("------Add Friend------");
 		return axios.post(`/api/list/${user}`, friend);
 	}
 }
